Добавить возможность видеть активные задачи, над которыми работают сотрудники в настоящий момент

+++ app/views/user/dashboard.blade.php
@@ -4,9 +4,9 @@
 @section('content')
 <?php
     $tasks = array();
-    $projects = Project::where('superior_id',Auth::user()->id)->orderBy('updated_at','DESC')->with('team','tasks')->limit(4)->get();
-    if($projects->count()):
-        $projectsIDs = Project::where('superior_id',Auth::user()->id)->lists('id');
+    $projects = ProjectOwners::where('user_id',Auth::user()->id)->orderBy('updated_at','DESC')->with('projects','projects.client','projects.team','projects.tasks')->limit(4)->get();
+    if(count($projects)):
+        $projectsIDs = ProjectOwners::where('user_id',Auth::user()->id)->lists('project_id');
         $tasks = ProjectTask::whereIn('project_id',$projectsIDs)->where('start_status',1)->where('stop_status',0)->with('cooperator','project')->get();
     endif;
     $dt_request = Request::get('date') ? Request::get('date') : date('Y-m-d');
@@ -24,48 +24,51 @@
 <div class="row placeholders">
     @foreach($projects as $project)
         <div class="col-xs-6 col-sm-3 placeholder">
-            <a href="{{ URL::route('projects.show',$project->id) }}" class="">
+            <a href="{{ URL::route('projects.show',$project->projects->id) }}" class="">
                 <img data-src="holder.js/200x200/auto/sky" class="img-responsive" alt="Generic placeholder thumbnail">
             </a>
-            <a href="{{ URL::route('projects.show',$project->id) }}" class=""><h4>{{ $project->title }}</h4></a>
-            <span class="text-muted">{{ $project->description }}</span>
-            @if($project->team->count())
-                <br><span class="text-muted">{{ $project->team->count() }} {{ Lang::choice('участник|участника|участников',$project->team->count()) }}</span>
+            <a href="{{ URL::route('projects.show',$project->projects->id) }}" class=""><h4>{{ $project->projects->title }}</h4></a>
+            <span class="text-muted">{{ $project->projects->description }}</span>
+            @if(count($project->projects->team))
+                <br><span class="text-muted">{{ $project->projects->team->count() }} {{ Lang::choice('участник|участника|участников',$project->projects->team->count()) }}</span>
             @endif
-            @if($project->tasks->count())
-                <br><span class="text-muted">{{ $project->tasks->count() }} {{ Lang::choice('задача|задачи|задач',$project->tasks->count()) }}</span>
+            @if(count($project->projects->tasks))
+                <br><span class="text-muted">{{ $project->projects->tasks->count() }} {{ Lang::choice('задача|задачи|задач',$project->projects->tasks->count()) }}</span>
             @endif
         </div>
     @endforeach
 </div>
 @else
-
+    <p>В не являететсь админитратором проектов</p>
 @endif
 @if(count($tasks))
 <h2 class="sub-header">Список активных задач</h2>
 <div class="table-responsive">
     <table class="table table-striped">
         <tbody>
+        <?php $tasks_total_time = 0;?>
+        <?php $tasks_total_price = 0;?>
         <?php $earnMoneyCurrentDate = costCalculation(NULL,['tasks' => $tasks]);?>
-        <?php $earnMoneyTotal = costCalculation();?>
         @foreach($tasks as $task)
+            <?php $tasks_total_time += (getLeadTimeMinutes($task)+floor($task->lead_time/60));?>
+            @if(isset($earnMoneyCurrentDate[$task->id]['earnings']))
+                <?php $tasks_total_price += $earnMoneyCurrentDate[$task->id]['earnings'];?>
+            @endif
             <tr>
                 <td>
                     <a href="{{ URL::route('projects.show',$task->project->id) }}">{{ $task->project->title }}</a>
                     <br>
-                    <a href="{{ URL::route('cooperators.show',$task->cooperator->id) }}">{{ getInitials($task->cooperator->fio) }}</a>
-                    <br>
-                    {{ $task->note }}
-                </td>
-                <td>
-                    {{ culcLeadTime($task) }} / {{ isset($earnMoneyCurrentDate[$task->project->id][$task->id]['earnings']) ? number_format($earnMoneyCurrentDate[$task->project->id][$task->id]['earnings'],2,'.',' ').' руб.' : '' }}
-                    @if(isset($earnMoneyTotal[$task->project->id][$task->id]['overdose']) && $earnMoneyTotal[$task->project->id][$task->id]['overdose'] == 1)
-                        <br><span class="label label-danger">Превышен допустимый лимит бюджета</span>
-                        <br><span class="label label-info">Текущий заработок: {{ number_format($earnMoneyTotal[$task->project->id][$task->id]['overdose_money'],2,'.',' ').' руб.' }}</span>
-                        <br><span class="label label-info">Доступный бюджет: {{ number_format($earnMoneyTotal[$task->project->id][$task->id]['budget'],2,'.',' ').' руб.' }}</span>
+                    {{ getInitials($task->cooperator->fio) }}
+                    <br>{{ $task->note }}
+                    @if(count($task->basecamp_task))
+                        <a href="{{ $task->basecamp_task->basecamp_task_link  }}" target="_blank"><span aria-hidden="true" class="glyphicon glyphicon-new-window"></span></a>
                     @endif
                 </td>
                 <td>
+                    {{ culcLeadTime($task) }} @if($task->user_id == Auth::user()->id)/ {{ isset($earnMoneyCurrentDate[$task->id]['earnings']) ? number_format($earnMoneyCurrentDate[$task->id]['earnings'],2,'.',' ').' руб.' : '' }}@endif
+                </td>
+                <td>
+                @if($task->user_id == Auth::user()->id)
                     {{ Form::open(array('route'=>array('timesheets.run_timer'),'method'=>'POST','style'=>'display:inline-block')) }}
                     {{ Form::hidden('task',$task->id) }}
                     @if($task->start_status && !$task->stop_status)
@@ -76,18 +79,26 @@
                         {{ Form::submit('Продолжить',['class'=>'btn btn-default']) }}
                     @endif
                     {{ Form::close() }}
+                @endif
                 </td>
-                <td><a href="{{ URL::route('timesheets.edit',[$task->id,'date'=>$dt_request]) }}" class="btn btn-success">Редактировать</a></td>
                 <td>
+                @if($task->user_id == Auth::user()->id)
+                    <a href="{{ URL::route('timesheets.edit',[$task->id,'date'=>$dt_request]) }}" class="btn btn-success">Редактировать</a></td>
+                @endif
+                <td>
+                @if($task->user_id == Auth::user()->id)
                     {{ Form::open(array('route'=>array('timesheets.destroy',$task->id),'method'=>'DELETE','style'=>'display:inline-block')) }}
                     {{ Form::submit('Удалить',['class'=>'btn btn-danger']) }}
                     {{ Form::close() }}
+                @endif
                 </td>
             </tr>
         @endforeach
         </tbody>
     </table>
 </div>
+@else
+    <p>Сейчас никто не выполняет задачи</p>
 @endif
 @stop
 @section('scripts')
